+++ src/components/UI/Display/Table/Table.tsx
@@ -1,0 +1,3 @@
+import { Table } from 'grommet';
+
+export default Table;
