+++ src/components/UI/Display/Table/index.tsx
@@ -1,3 +1,8 @@
-import { Table } from 'grommet';
+import Table from './Table';
+import TableBody from './TableBody';
+import TableCell from './TableCell';
+import TableFooter from './TableFooter';
+import TableHeader from './TableHeader';
+import TableRow from './TableRow';
 
-export default Table;
+export { Table, TableBody, TableCell, TableFooter, TableHeader, TableRow };
